refactor: extract config roundtrip detector
